Update SetupGuideLayout to deliver either cloud or self-managed instructions
